Do the happy requests per second!

+++ hystrix-sample/src/main/java/io/acari/IdRepository.java
@@ -9,6 +9,6 @@ public class IdRepository {
     private final SecureRandom secureRandom = new SecureRandom();
 
     public Long getRanbo(){
-        return secureRandom.nextLong();
+        return (long) secureRandom.nextInt();
     }
 }
